Fix old bugs

+++ veche-web/src/Handler/Issue.hs
@@ -137,7 +137,7 @@ makePollWidget mPoll issueId IssueMaterialized{isVoteAllowed, votes} =
                 sum [ Map.findWithDefault 0 key weights
                     | User{stellarAddress = key} <- toList users
                     ]
-            percentage = realToFrac (choiceWeight / sum weights * 100) :: Double
+            percentage = choiceWeight / sum weights * 100 :: Double
             share =
                 printf
                     "%d/%d"
